moved tests to different folder
Moved tests to different folder.
