+++ jest.config.ts
@@ -8,9 +8,9 @@ export default async (): Promise<Config> => {
     collectCoverage:true,
     // testMatch: ['**/+(*.)+(unit|test).+(ts|js)?(x)'],
     // collectCoverageFrom: ['./src/**/*.{ts,js}','./lib/**/*.{ts,js}'],
+    testMatch: ["**/test/*.ts", "**/?(*.)+(spec|test).[tj]s?(x)"],
     transform: {
-      '^.+\\.(ts|tsx)?$': 'ts-jest',
-      "^.+\\.(js|jsx)$": "babel-jest",
+      '^.+\\.(ts|tsx)?$': 'ts-jest'
     }
 
   };
