+++ tsconfig.json
@@ -25,7 +25,7 @@
     // "moduleDetection": "auto",                        /* Control what method is used to detect module-format JS files. */
     /* Modules */
     "module": "commonjs", /* Specify what module code is generated. */
-    "rootDir": "src", /* Specify the root folder within your source files. */
+    // "rootDir": "src", /* Specify the root folder within your source files. */
     "moduleResolution": "node", /* Specify how TypeScript looks up a file from a given module specifier. */
     // "baseUrl": ".",                                  /* Specify the base directory to resolve non-relative module names. */
     // "paths": {
@@ -33,7 +33,10 @@
     //   "@core/model":["./src/core/model"],
     //   "@core/queue":["./src/core/queue"]
     // },                                      /* Specify a set of entries that re-map imports to additional lookup locations. */
-    // "rootDirs": [],                                   /* Allow multiple folders to be treated as one when resolving modules. */
+    "rootDirs": [
+      "src",
+      "test"
+    ],                                   /* Allow multiple folders to be treated as one when resolving modules. */
     // "typeRoots": [],                                  /* Specify multiple folders that act like './node_modules/@types'. */
     // "types": [],                                      /* Specify type package names to be included without being referenced in a source file. */
     // "allowUmdGlobalAccess": true,                     /* Allow accessing UMD globals from modules. */
@@ -102,8 +105,8 @@
     "node_modules",
     "lib",
     "coverage",
-    "src/**/*.unit.ts",
     "*.config.ts",
-    "*.config.js"
+    "*.config.js",
+    "test"
   ]
 }
